Dynamically import and upgrade Monaco
