+++ web/src/languages/asm.ts
@@ -1,4 +1,4 @@
-import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
+import type * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 import { base } from "./base";
 
 export const AsmLanguage: monaco.languages.IMonarchLanguage = {
