+++ web/src/languages/base.ts
@@ -1,4 +1,4 @@
-import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
+import type * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 export const base: monaco.languages.IMonarchLanguage = {
   tokenizer: {
     comment: [
