+++ web/src/languages/cmp.ts
@@ -1,4 +1,4 @@
-import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
+import type * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 
 export const CmpLanguage: monaco.languages.IMonarchLanguage = {
   tokenizer: {
