+++ web/src/languages/vm.ts
@@ -1,4 +1,4 @@
-import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
+import type * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 import { base } from "./base";
 
 export const VmLanguage: monaco.languages.IMonarchLanguage = {
